requirements.txt
elasticsearch==8.14.0
langchain_community==0.2.6
langchain_core==0.2.11
langchain_text_splitters==0.2.2
python-dotenv==1.0.1
